Break out classes into Smab.DebugInfo.Models
