+++ src/Smab.DebugInfo/Models/MVCStructure.MVCController.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections.Generic;
+
+namespace Smab.DebugInfo.Models
+{
+	public partial class MVCStructure
+	{
+		public class MVCController
+		{
+			public string Name { get; set; } = "";
+			public List<string> Fields = new List<string>();
+			public List<MVCAction> Actions = new List<MVCAction>();
+		}
+	}
+
+}
