+++ src/Smab.DebugInfo/Models/MVCStructure.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Smab.DebugInfo.Models
+{
+	public partial class MVCStructure
+	{
+		public List<MVCController> Controllers { get; set; } = new List<MVCController>();
+	}
+
+}
