+++ src/Smab.DebugInfo/Smab.DebugInfo.csproj
@@ -14,12 +14,12 @@
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Authors>Simon Brookes (@smabuk)</Authors>
     <Company></Company>
-    <Description>Diagnostic middleware for ASP.NET Core to display information useful when debugging an ASP.NET Core 2.x running environment.</Description>
-    <Summary>Diagnostic middleware for ASP.NET Core to display information useful when debugging an ASP.NET Core 2.x running environment.</Summary>
+    <Description>Diagnostic middleware for ASP.NET Core to display information useful when debugging an ASP.NET Core 2.x or 3.x running environment.</Description>
+    <Summary>Diagnostic middleware for ASP.NET Core to display information useful when debugging an ASP.NET Core 2.x or 3.x running environment.</Summary>
     <PackageProjectUrl>https://github.com/smabuk/Smab.DebugInfo</PackageProjectUrl>
     <RepositoryUrl>https://github.com/smabuk/DebugInfo</RepositoryUrl>
     <PackageTags>asp.net core, diagnostic, assemblies, configuration, config</PackageTags>
-    <PackageReleaseNotes>Switch to System.Text.Json</PackageReleaseNotes>
+    <PackageReleaseNotes>Break out classes into Smab.DebugInfo.Models</PackageReleaseNotes>
     <VersionPrefix>0.9.8</VersionPrefix>
     <Preview>preview</Preview>
     <VersionSuffix Condition="'$(Preview)' != ''">$(Preview).$([System.DateTime]::get_Now().get_Year())$([System.DateTime]::get_Now().get_Month().ToString("D2"))$([System.DateTime]::get_Now().get_Day().ToString("D2"))-$([System.DateTime]::get_Now().get_Hour().ToString("D2"))$([System.DateTime]::get_Now().get_Minute().ToString("D2"))</VersionSuffix>
